Change particle positions, velocities and tree node position and size alongside some variables to double. Seems to fix noise issues where particles ended up in wrong nodes because of rounding error to 32 bit.

+++ src/globals.h
@@ -143,8 +143,8 @@ extern struct ImageProperties {
 
 
 extern struct ParticleData {
-    float Pos[3];
-    float Vel[3];
+    double Pos[3];
+    double Vel[3];
     int32_t ID;
     int Type;
     peanoKey Key;
